PHP8.1 - Fix implicit conversion to int

+++ src/PHPExcelReader/SpreadsheetReader.php
@@ -723,7 +723,7 @@ class SpreadsheetReader
             $this->setOutputEncoding($outputEncoding);
         }
         for ($i = 1; $i < 245; $i++) {
-            $name = strtolower(((($i - 1) / 26 >= 1) ? chr(($i - 1) / 26 + 64) : '') . chr(($i - 1) % 26 + 65));
+            $name = strtolower(((($i - 1) / 26 >= 1) ? chr((int)(($i - 1) / 26 + 64)) : '') . chr(($i - 1) % 26 + 65));
             $this->colnames[$name] = $i;
             $this->colindexes[$i] = $name;
         }
